Add sheell completion for names in solution command
