+++ cmd/solution/status.go
@@ -56,6 +56,9 @@ var solutionStatusCmd = &cobra.Command{
 		}
 	},
 	TraverseChildren: true,
+	ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		return getSolutionNames(toComplete), cobra.ShellCompDirectiveDefault
+	},
 }
 
 func getSolutionStatusCmd() *cobra.Command {
