+++ cmd/solution/unsubscribe.go
@@ -32,6 +32,9 @@ var solutionUnsubscribeCmd = &cobra.Command{
 	Example:          `  fsoc solution unsubscribe spacefleet`,
 	Run:              unsubscribeFromSolution,
 	TraverseChildren: true,
+	ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		return getSolutionNames(toComplete), cobra.ShellCompDirectiveDefault
+	},
 }
 
 func getUnsubscribeSolutionCmd() *cobra.Command {
